Add support for template being an array
There are cases where you want the response from a service to be an
array, but the collection support doesn't work. For instance, if you
want things in a specific order, or you want the number of things to
depend on some query parameter (e.g. "?limit=10").

With this patch, it's possible to define the template property of a
service as an array (or a function returning an array), for instance:

    template: function(params, query) {
        var things = [];
        for (var i = 0; i < query.limit; i++) {
            things.push({ foo: params.name, id: i + 1 });
        }
        return things;
    }